\name{newFocus-package}
\alias{newFocus-package}
\docType{package}
\title{
\packageTitle{newFocus}
}
\description{
\packageDescription{newFocus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{newFocus}
\packageIndices{newFocus}
For the GO (Gene Ontology) terms chosen as focus level nodes, \link{newFocus} function will return the minimum number of true discoveries. For GO terms that are non-focus level nodes, we use \link{pick} to count the number of true discoveries based on the result of \link{newFocus}.   
}
\author{
\packageAuthor{newFocus}

Maintainer: \packageMaintainer{newFocus}
}
\references{
  Ningning Xu, Aldo solari, Jelle Goeman, Clsoed testing with global test, with applications on metabolomics data, arXiv:2001.01541, https://arxiv.org/abs/2001.01541
  
  Jelle J. Goeman, Sara A. van de Geer, Floor de Kort, Hans C. van Houwelingen, A global test for groups of genes: testing association with a clinical outcome, Bioinformatics, Volume 20, Issue 1, 1 January 2004, Pages 93-99, https://doi.org/10.1093/bioinformatics/btg382
}

