\name{newFocus-package}
\alias{newFocus-package}
\docType{package}
\title{
\packageTitle{newFocus}
}
\description{
\packageDescription{newFocus}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{newFocus}
\packageIndices{newFocus}
For the GO (Gene Ontology) terms chosen as focus level nodes, \link{newFocus} function will return the number. For GO terms that are non-focus level nodes, we use \link{pick} to count the number of true discoveries based on the result of \link{newFocus}.  
}
\author{
\packageAuthor{newFocus}

Maintainer: \packageMaintainer{newFocus}
}

